major restructorization, redesigned search, cityinfo, language switcher
