+++ source/js/components/Dashboard/Sidebar/Filter/CategoryFilterItem.jsx
