+++ source/js/components/Dashboard/Sidebar/ResultList/ResultListItem.jsx
