+++ source/js/config/config.js
@@ -9,4 +9,4 @@ export const LANGUAGE_CODE = 'de_AT';
 export const COUNTRY_CODE = 'at';
 export const MAPBOX_API_KEY = 'pk.eyJ1Ijoid2lraW1lZGlhYXVzdHJpYSIsImEiOiJjajUyZWw3MTMwZGVpMnJxbGsxdnN4ZnlqIn0.9D5QSHyd8fs0_Wg2zlFnQg';
 
-export const WIKIDAHEIM_ENDPOINT = 'https://wikidaheim.at/api/api.php';
+export const WIKIDAHEIM_ENDPOINT = 'https://api.wikidaheim.at/api.php';
