+++ source/js/reducers/app.js
@@ -110,17 +110,10 @@ const actionsMap = {
   },
 
   [PLACE_SELECT_ACTION_START]: (state, action) => {
-    /* add the region to the municipality (only available when triggered fron search) */
-    const region = action.data.get('context') ? action.data.get('context').find((context) => {
-      const id = context.get('id').split('.');
-      if (id[0] === 'region') return true;
-      return false;
-    }) : false;
-
     return state.merge({
       searchLoading: false,
       searchData: new List(),
-      searchText: region ? `${ action.data.get('text') }, ${ region.get('text') }` : action.data.get('text'),
+      searchText: action.data.get('text'),
       placeMapData: action.data,
       placeLoading: true,
       placeError: null,
